splits temp into header and cpp
